requirements.txt
streamlit==1.10.0
numpy==1.20.3
Pillow==8.2.0
